+++ src/java/org/apache/fop/render/java2d/Java2DGraphics2DAdapter.java
@@ -51,8 +51,8 @@ public class Java2DGraphics2DAdapter extends AbstractGraphics2DAdapter {
         float imw = (float)dim.getWidth() / 1000f;
         float imh = (float)dim.getHeight() / 1000f;
 
-        float sx = fwidth / (float)imw;
-        float sy = fheight / (float)imh;
+        float sx = fwidth / imw;
+        float sy = fheight / imh;
 
         Java2DRenderer renderer = (Java2DRenderer)context.getRenderer();
         Java2DGraphicsState state = renderer.state;
